refactor(backend): extract accounts email and lifecycle helpers from views
Thin-views treatment for the accounts app (previously untouched by the
refactoring campaign): the module-level helper layer moves out of
accounts/views.py (615 -> 424 lines) into two modules:

- accounts/emails.py: activation / password-reset / email-change
  delivery, frontend token-link building and the local-email-backend
  check. Tests now patch accounts.emails.send_mail (3 targets
  repointed).
- accounts/services.py: activation-expiry bookkeeping (with
  ACTIVATION_EXPIRY_DAYS), session invalidation, uid decoding, email
  normalization and German-locale serializer validation.

All bodies are moved verbatim; the German response-message constants
and _registration_success_message stay in views.py because they are
response concerns. No behavior change.
